first commit of `SiameseTextClassifier` class
Implements the `SiameseTextClassifier` class, which extends the
`TextClassifier` class by implementing a Siamese architecture with twin
networks of RNN, CNN, and/or simple dense layers that share weights
across the two sub-networks (but up/down layers).

This architecture is useful for cases in which inputs are pairs of examples (e.g.  classifying name1 and name2 as a match or not).